feat adicionando calculo parte 1

+++ src/main/java/br/csi/projeto_calculo_racao/model/tipoRacao/TipoRacaoRepository.java
@@ -2,5 +2,5 @@ package br.csi.projeto_calculo_racao.model.tipoRacao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoRacaoRepository extends JpaRepository<TipoRacao, Integer> {
+public interface TipoRacaoRepository extends JpaRepository<TipoRacao, Long> {
 }
